Have added the rocket component inside the RocketList component
